vala_0_38, vala_0_40, vala_0_42: add configuration to disable graphviz
This allows building Vala without support for Graphviz; useful for more
minimal installs where we don't want to pull it (and transitively,
pango, gd, etc.) in as a dependency.

NOTE: this is a slightly modified version of the upstream PR I submitted
that uses `if/else` instead of `lib.optional[s]`, to avoid having to
rebuild everything that depends on vala.